requirements.txt
ffmpeg-python
Pillow
python-magic
